refactor: shampoo_utils
